Resources customization commit

+++ src/main/java/com/spingular/cms/web/rest/ActivityResource.java
@@ -1,28 +1,41 @@
 package com.spingular.cms.web.rest;
 
-import com.spingular.cms.repository.ActivityRepository;
-import com.spingular.cms.service.ActivityQueryService;
-import com.spingular.cms.service.ActivityService;
-import com.spingular.cms.service.criteria.ActivityCriteria;
-import com.spingular.cms.service.dto.ActivityDTO;
-import com.spingular.cms.web.rest.errors.BadRequestAlertException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
+
+import com.spingular.cms.repository.ActivityRepository;
+import com.spingular.cms.security.AuthoritiesConstants;
+import com.spingular.cms.security.SecurityUtils;
+import com.spingular.cms.service.ActivityQueryService;
+import com.spingular.cms.service.ActivityService;
+import com.spingular.cms.service.criteria.ActivityCriteria;
+import com.spingular.cms.service.dto.ActivityDTO;
+import com.spingular.cms.web.rest.errors.BadRequestAlertException;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
 import tech.jhipster.web.util.HeaderUtil;
 import tech.jhipster.web.util.PaginationUtil;
 import tech.jhipster.web.util.ResponseUtil;
@@ -47,11 +60,8 @@ public class ActivityResource {
 
     private final ActivityQueryService activityQueryService;
 
-    public ActivityResource(
-        ActivityService activityService,
-        ActivityRepository activityRepository,
-        ActivityQueryService activityQueryService
-    ) {
+    public ActivityResource(ActivityService activityService, ActivityRepository activityRepository,
+            ActivityQueryService activityQueryService) {
         this.activityService = activityService;
         this.activityRepository = activityRepository;
         this.activityQueryService = activityQueryService;
@@ -61,37 +71,47 @@ public class ActivityResource {
      * {@code POST  /activities} : Create a new activity.
      *
      * @param activityDTO the activityDTO to create.
-     * @return the {@link ResponseEntity} with status {@code 201 (Created)} and with body the new activityDTO, or with status {@code 400 (Bad Request)} if the activity has already an ID.
+     * @return the {@link ResponseEntity} with status {@code 201 (Created)} and with
+     *         body the new activityDTO, or with status {@code 400 (Bad Request)} if
+     *         the activity has already an ID.
      * @throws URISyntaxException if the Location URI syntax is incorrect.
      */
     @PostMapping("/activities")
-    public ResponseEntity<ActivityDTO> createActivity(@Valid @RequestBody ActivityDTO activityDTO) throws URISyntaxException {
+    public ResponseEntity<ActivityDTO> createActivity(@Valid @RequestBody ActivityDTO activityDTO)
+            throws URISyntaxException {
         log.debug("REST request to save Activity : {}", activityDTO);
         if (activityDTO.getId() != null) {
             throw new BadRequestAlertException("A new activity cannot already have an ID", ENTITY_NAME, "idexists");
         }
-        ActivityDTO result = activityService.save(activityDTO);
+
+        // NOTE: If anyone can use this Object, anyone can Create & Read any object
+        ActivityDTO result = new ActivityDTO();
+        if (SecurityUtils.hasCurrentUserThisAuthority(AuthoritiesConstants.USER)
+                || SecurityUtils.hasCurrentUserThisAuthority(AuthoritiesConstants.ADMIN)) {
+            result = activityService.save(activityDTO);
+        }
+
         return ResponseEntity
-            .created(new URI("/api/activities/" + result.getId()))
-            .headers(HeaderUtil.createEntityCreationAlert(applicationName, true, ENTITY_NAME, result.getId().toString()))
-            .body(result);
+                .created(new URI("/api/activities/" + result.getId())).headers(HeaderUtil
+                        .createEntityCreationAlert(applicationName, true, ENTITY_NAME, result.getId().toString()))
+                .body(result);
     }
 
     /**
      * {@code PUT  /activities/:id} : Updates an existing activity.
      *
-     * @param id the id of the activityDTO to save.
+     * @param id          the id of the activityDTO to save.
      * @param activityDTO the activityDTO to update.
-     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and with body the updated activityDTO,
-     * or with status {@code 400 (Bad Request)} if the activityDTO is not valid,
-     * or with status {@code 500 (Internal Server Error)} if the activityDTO couldn't be updated.
+     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and with body
+     *         the updated activityDTO, or with status {@code 400 (Bad Request)} if
+     *         the activityDTO is not valid, or with status
+     *         {@code 500 (Internal Server Error)} if the activityDTO couldn't be
+     *         updated.
      * @throws URISyntaxException if the Location URI syntax is incorrect.
      */
     @PutMapping("/activities/{id}")
-    public ResponseEntity<ActivityDTO> updateActivity(
-        @PathVariable(value = "id", required = false) final Long id,
-        @Valid @RequestBody ActivityDTO activityDTO
-    ) throws URISyntaxException {
+    public ResponseEntity<ActivityDTO> updateActivity(@PathVariable(value = "id", required = false) final Long id,
+            @Valid @RequestBody ActivityDTO activityDTO) throws URISyntaxException {
         log.debug("REST request to update Activity : {}, {}", id, activityDTO);
         if (activityDTO.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
@@ -104,29 +124,35 @@ public class ActivityResource {
             throw new BadRequestAlertException("Entity not found", ENTITY_NAME, "idnotfound");
         }
 
-        ActivityDTO result = activityService.save(activityDTO);
-        return ResponseEntity
-            .ok()
-            .headers(HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, activityDTO.getId().toString()))
-            .body(result);
+       // NOTE: Only admins can change an Object (that belongs to everyone)
+        ActivityDTO result = new ActivityDTO();
+        if (SecurityUtils.hasCurrentUserThisAuthority(AuthoritiesConstants.ADMIN)) {
+            result = activityService.save(activityDTO);
+        }
+
+        return ResponseEntity.ok().headers(
+                HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, activityDTO.getId().toString()))
+                .body(result);
     }
 
     /**
-     * {@code PATCH  /activities/:id} : Partial updates given fields of an existing activity, field will ignore if it is null
+     * {@code PATCH  /activities/:id} : Partial updates given fields of an existing
+     * activity, field will ignore if it is null
      *
-     * @param id the id of the activityDTO to save.
+     * @param id          the id of the activityDTO to save.
      * @param activityDTO the activityDTO to update.
-     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and with body the updated activityDTO,
-     * or with status {@code 400 (Bad Request)} if the activityDTO is not valid,
-     * or with status {@code 404 (Not Found)} if the activityDTO is not found,
-     * or with status {@code 500 (Internal Server Error)} if the activityDTO couldn't be updated.
+     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and with body
+     *         the updated activityDTO, or with status {@code 400 (Bad Request)} if
+     *         the activityDTO is not valid, or with status {@code 404 (Not Found)}
+     *         if the activityDTO is not found, or with status
+     *         {@code 500 (Internal Server Error)} if the activityDTO couldn't be
+     *         updated.
      * @throws URISyntaxException if the Location URI syntax is incorrect.
      */
     @PatchMapping(value = "/activities/{id}", consumes = "application/merge-patch+json")
     public ResponseEntity<ActivityDTO> partialUpdateActivity(
-        @PathVariable(value = "id", required = false) final Long id,
-        @NotNull @RequestBody ActivityDTO activityDTO
-    ) throws URISyntaxException {
+            @PathVariable(value = "id", required = false) final Long id, @NotNull @RequestBody ActivityDTO activityDTO)
+            throws URISyntaxException {
         log.debug("REST request to partial update Activity partially : {}, {}", id, activityDTO);
         if (activityDTO.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
@@ -139,12 +165,16 @@ public class ActivityResource {
             throw new BadRequestAlertException("Entity not found", ENTITY_NAME, "idnotfound");
         }
 
+       // NOTE: Only admins can change an Object (that belongs to everyone)
         Optional<ActivityDTO> result = activityService.partialUpdate(activityDTO);
+        if (SecurityUtils.hasCurrentUserThisAuthority(AuthoritiesConstants.ADMIN)) {
+            if (result.isPresent()) {
+                activityService.save(activityDTO);
+            }
+        }
 
-        return ResponseUtil.wrapOrNotFound(
-            result,
-            HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, activityDTO.getId().toString())
-        );
+        return ResponseUtil.wrapOrNotFound(result,
+                HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, activityDTO.getId().toString()));
     }
 
     /**
@@ -152,13 +182,18 @@ public class ActivityResource {
      *
      * @param pageable the pagination information.
      * @param criteria the criteria which the requested entities should match.
-     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and the list of activities in body.
+     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and the list
+     *         of activities in body.
      */
     @GetMapping("/activities")
     public ResponseEntity<List<ActivityDTO>> getAllActivities(ActivityCriteria criteria, Pageable pageable) {
         log.debug("REST request to get Activities by criteria: {}", criteria);
+
+        // NOTE: If anyone can use this Object, anyone can Create & Read any object
         Page<ActivityDTO> page = activityQueryService.findByCriteria(criteria, pageable);
-        HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(ServletUriComponentsBuilder.fromCurrentRequest(), page);
+
+        HttpHeaders headers = PaginationUtil
+                .generatePaginationHttpHeaders(ServletUriComponentsBuilder.fromCurrentRequest(), page);
         return ResponseEntity.ok().headers(headers).body(page.getContent());
     }
 
@@ -166,7 +201,8 @@ public class ActivityResource {
      * {@code GET  /activities/count} : count all the activities.
      *
      * @param criteria the criteria which the requested entities should match.
-     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and the count in body.
+     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and the count
+     *         in body.
      */
     @GetMapping("/activities/count")
     public ResponseEntity<Long> countActivities(ActivityCriteria criteria) {
@@ -178,13 +214,18 @@ public class ActivityResource {
      * {@code GET  /activities/:id} : get the "id" activity.
      *
      * @param id the id of the activityDTO to retrieve.
-     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and with body the activityDTO, or with status {@code 404 (Not Found)}.
+     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and with body
+     *         the activityDTO, or with status {@code 404 (Not Found)}.
      */
     @GetMapping("/activities/{id}")
     public ResponseEntity<ActivityDTO> getActivity(@PathVariable Long id) {
         log.debug("REST request to get Activity : {}", id);
+
+        // NOTE: If anyone can use this Object, anyone can Create & Read any object
         Optional<ActivityDTO> activityDTO = activityService.findOne(id);
+
         return ResponseUtil.wrapOrNotFound(activityDTO);
+
     }
 
     /**
@@ -196,10 +237,14 @@ public class ActivityResource {
     @DeleteMapping("/activities/{id}")
     public ResponseEntity<Void> deleteActivity(@PathVariable Long id) {
         log.debug("REST request to delete Activity : {}", id);
-        activityService.delete(id);
-        return ResponseEntity
-            .noContent()
-            .headers(HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString()))
-            .build();
+
+        // NOTE: Only admins can delete an Object (that belongs to everyone)
+        if (SecurityUtils.hasCurrentUserThisAuthority(AuthoritiesConstants.ADMIN)) {
+            activityService.delete(id);
+        }
+
+        return ResponseEntity.noContent()
+                .headers(HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString()))
+                .build();
     }
 }
